Added collision boxes in the corners
