+++ Bomberman/Assets/Materials/Color.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5df27cee962302a408658c0248c975aa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
