+++ Bomberman/Assets/Materials/Wall.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6adfbfde0f65a74fac946b603ab355b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
