--- Bomberman/Assets/Movement_physics.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 887a9d959d187e1449295fd791a0db4b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
