+++ Bomberman/Assets/Player_Code.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 143b5362ddfb8e8469f5eff61f9c4a9d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
