+++ Bomberman/Assets/ProBuilder Data/ProBuilderPreferences.asset
@@ -11,22 +11,15 @@ MonoBehaviour:
   m_Script: {fileID: 680703436, guid: 4f0627da958b4bb78c260446066f065f, type: 3}
   m_Name: ProBuilderPreferences
   m_EditorClassIdentifier: 
-  m_Bool_keys:
-  - pbIconGUI
-  m_Int_keys:
-  - pbHandleAlignment
-  - pbDefaultEditLevel
-  m_Float_keys:
-  - pbEditorScroll.x
-  - pbEditorScroll.y
+  m_Bool_keys: []
+  m_Int_keys: []
+  m_Float_keys: []
   m_String_keys: []
   m_Color_keys: []
   m_Material_keys: []
-  m_Bool_values: 01
-  m_Int_values: 0000000004000000
-  m_Float_values:
-  - 0
-  - 0
+  m_Bool_values: 
+  m_Int_values: 
+  m_Float_values: []
   m_String_values: []
   m_Color_values: []
   m_Material_values: []
